requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cfgv==3.4.0
distlib==0.3.9
Django==4.2.16
django-cors-headers==4.4.0
django-currentuser==0.8.0
djangorestframework==3.15.2
factory_boy==3.3.1
Faker==30.6.0
filelock==3.16.1
identify==2.6.1
nodeenv==1.9.1
pillow==10.4.0
platformdirs==4.3.6
pre-commit==3.5.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
virtualenv==20.26.6
